Made necessary changes to get quest notifications working correctly
So far I have had to rewrite some of the questlog system due to it automatically adding all of the quests at the start of the game.

I have changed removed this and instead added a function that allows a quest to be passed in and added to the questlog. The quest is then removed from the questlist.

This will work a lot better with the quest notifications.

I've also moved the quest notification code back to the questlog script to keep it all together. In addition to this, I have added a queue for conversations if the NPC's talk more than once to the player. Along with this I have added a way to dequeue conversations.